gateway_client/requirements.txt
flask>=3.0.0
schedule>=1.2.0
requests>=2.31.0
python-dotenv>=1.0.0
pycryptodome>=3.20.0